requirements.txt
# BME Agent Requirements
# Core dependencies for Streamlit deployment
streamlit
python-dotenv
mistralai
anthropic
toml
requests
psycopg2-binary
python-dateutil
pytz

# To install: pip install -r requirements.txt
# For deployment: Ensure all packages are installed in your environment